+++ example/ec2/defaults.rb
@@ -1,0 +1,6 @@
+region 'us-east-2'
+instance_type 'm1.large'
+security_groups ['group1','group2']
+image_id 'ami-1234567'
+iam_instance_profile 'some_iam_profile'
+key_name 'YourKeyName'