+++ lib/rivet/as/launch_config.rb
@@ -5,18 +5,11 @@ module Rivet
 
     ATTRIBUTES = [
       :bootstrap,
-      :iam_instance_profile,
-      :image_id,
-      :instance_type,
-      :key_name,
-      :security_groups,
-      :associate_public_ip_address,
       :detailed_instance_monitoring,
-      :block_device_mappings,
-      :kernel_id,
-      :ramdisk_id,
       :spot_price
-    ].each { |a| attr_reader a }
+    ] + BASE_AWS_ATTRIBUTES
+
+    ATTRIBUTES.each { |a| attr_reader a }
 
     attr_reader :id_prefix, :config
 
@@ -71,7 +64,11 @@ module Rivet
     def build_identity_string
       identity = ATTRIBUTES.inject('') do |accum, attribute|
         if attribute != :bootstrap
-          attr_value = self.send(attribute) ? self.send(attribute) : "\0"
+          attr_value = if self.send(attribute).nil?
+            "\0"
+          else
+            self.send(attribute)
+          end
           attr_value = attr_value.join("\t")  if attr_value.respond_to? :join
           attr_value = attr_value.to_s        if !!attr_value == attr_value
           accum << attribute.to_s